trading-platform/symbolic_trading/requirements.txt
# Core Dependencies
numpy>=1.21.0
sympy>=1.9
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0
pytest-mock>=3.6.1

# API and Networking
aiohttp>=3.8.0
requests>=2.26.0

# Development
black>=21.9b0
flake8>=3.9.0

# Database
sqlalchemy>=1.4.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3
pyyaml>=5.4.1
